+++ precision/longd.h
@@ -6,7 +6,7 @@
 /*   By: wgorold <wgorold@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/05/20 18:19:18 by wgorold           #+#    #+#             */
-/*   Updated: 2019/06/05 19:19:42 by wgorold          ###   ########.fr       */
+/*   Updated: 2019/06/05 20:10:52 by wgorold          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -110,8 +110,8 @@ int		ft_putstr(char const *s);
 int		ft_putstrn(char const *s, int len);
 void	print_arr(short *list, int size);
 void	ft_putunbr(unsigned long n);
-void	ft_baseitoa(char *out, unsigned long target, unsigned char base, char big);
-void	ft_baseitoasign(char *out, long target, unsigned char base);
+void	ft_baseitoa(char *out, unsigned long long target, unsigned char base, char big);
+void	ft_baseitoasign(char *out, long long target, unsigned char base);
 
 void	init_t_str_f(t_str_f *input);
 void	init_t_str_f_ent(t_str_f *input);
